api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg[binary]==3.2.1
boto3==1.35.21
redis==5.0.8
rq==1.16.2
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
python-multipart==0.0.9
numpy==1.26.4
Pillow==10.4.0
ImageHash==4.3.1
insightface==0.7.3
onnxruntime==1.18.1
opencv-python-headless==4.10.0.84
psycopg_pool==3.2.1
pgvector==0.2.5
